requirements.txt
langchain_core
langchain_openai
langchain_community
langchain_google_genai
pypdf
langchain_text_splitters
langchain_chroma
arxiv
wikipedia
bs4
chroma-migrate
groq
langchain_groq
langchainhub
fastapi
uvicorn
langchain_google_community
google-api-python-client
# pysqlite3-binary
streamlit==1.28.1
chromadb==0.4.24
faiss-cpu